# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Um repositorio de autores e suas obras",
        "description": "Um repositorio para agrupar todos os autores e suas obras conhecidas da humanidade",
        "termsOfService": "https://minhaapi.com/terms",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/"
        }
    ],
    "paths": {
...
        "/Autor": {
            "get": {
                "tags": [
                    "Autor"
                ],
                "operationId": "1a9489842acc20e58065da6495c07a79",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Autor": {
                                                "description": "nome do Autor",
                                                "format": "string",
                                                "default": "tolkien"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Autor"
                ],
                "summary": "Criar um autor",
                "description": "Cria um autor",
                "operationId": "98b5ffd7c1d1c7b3aef5a21b76328601",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "description": "nome do autor",
                                            "format": "string",
                                            "default": "tokien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/Autor/{Autor}": {
            "get": {
                "tags": [
                    "Autor"
                ],
                "operationId": "6e205f136862cf08f4abc1042cf5de7d",
                "parameters": [
                    {
                        "name": "Autor",
                        "in": "path",
                        "description": "Id do Autor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "description": "tolkien",
                                            "format": "string",
                                            "default": "tolkien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Autor"
                ],
                "summary": "Atualizar um Autor",
                "description": "Atualizar um Autor",
                "operationId": "b9c0d64501d531bd87414d9635367f66",
                "parameters": [
                    {
                        "name": "Autor",
                        "in": "path",
                        "description": "Id do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "description": "nome do autor",
                                            "format": "string",
                                            "default": "tolkien"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Autor"
                ],
                "summary": "Remove o Autor especifico",
                "description": "Remove o livro",
                "operationId": "079ab20256d629507bb3b3d6c6427328",
                "parameters": [
                    {
                        "name": "Autor",
                        "in": "path",
                        "description": "Id do livro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "string": {
                                            "description": " string indicando a mensagem",
                                            "format": "string",
                                            "default": " msg"
                                        },
                                        "": {
                                            "description": "mensagem falando que o Autor foi deletado",
                                            "format": "string",
                                            "default": "Autor deletado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Pegue o token indo na operação login e passando o email e senha",
                "name": "Token JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}